Added testcases, new comments, changed title code
Had to fix title code if statement cause the title was always getting
set to an empty string. Made new test cases for new methods in
ApiDataRequest class and also added a new class in test.bingewolves and
that class helps start the JavaFX runtime thread so that my test can use
the JavaFX elements.